refactor: decompose table-view.tsx into focused sub-components (#685)
