+++ src/components/database/property-type-picker.test.ts
@@ -24,6 +24,13 @@ function readTableView(): string {
   );
 }
 
+function readTableCell(): string {
+  return readFileSync(
+    resolve(__dirname, "./views/table-cell.tsx"),
+    "utf-8",
+  );
+}
+
 function readPropertyTypePicker(): string {
   return readFileSync(
     resolve(__dirname, "./property-type-picker.tsx"),
@@ -98,21 +105,21 @@ describe("default column name matches property type (#563)", () => {
  * DateEditor (calendar picker) instead of a plain text input.
  */
 describe("table view uses registry editors for specialized types (#563)", () => {
-  const tableView = readTableView();
+  // Cell editing logic was extracted from table-view.tsx to table-cell.tsx
+  const tableCell = readTableCell();
 
   it("uses getPropertyTypeConfig to look up editors in the editing path", () => {
-    // The editing branch must call getPropertyTypeConfig to get the Editor
-    expect(tableView).toMatch(/getPropertyTypeConfig\(propertyType\)/);
+    expect(tableCell).toMatch(/getPropertyTypeConfig\(propertyType\)/);
   });
 
   it("renders a RegistryEditorCell for non-text-input types", () => {
-    expect(tableView).toContain("RegistryEditorCell");
+    expect(tableCell).toContain("RegistryEditorCell");
   });
 
   it("does not use plain input for date type editing", () => {
     // The plain-input exclusion list only includes text, number, url, email, phone.
     // Date is NOT excluded, so it uses the registry editor (DateEditor / calendar picker).
-    const plainInputTypes = tableView.match(
+    const plainInputTypes = tableCell.match(
       /hasRegistryEditor\s*=[\s\S]+?;/,
     )?.[0];
     expect(plainInputTypes).toBeDefined();
@@ -121,10 +128,10 @@ describe("table view uses registry editors for specialized types (#563)", () => 
   });
 
   it("RegistryEditorCell passes value, property, onChange, and onBlur to Editor", () => {
-    expect(tableView).toMatch(/<Editor[\s\S]*?value=\{/);
-    expect(tableView).toMatch(/<Editor[\s\S]*?property=\{property\}/);
-    expect(tableView).toMatch(/<Editor[\s\S]*?onChange=\{/);
-    expect(tableView).toMatch(/<Editor[\s\S]*?onBlur=\{/);
+    expect(tableCell).toMatch(/<Editor[\s\S]*?value=\{/);
+    expect(tableCell).toMatch(/<Editor[\s\S]*?property=\{property\}/);
+    expect(tableCell).toMatch(/<Editor[\s\S]*?onChange=\{/);
+    expect(tableCell).toMatch(/<Editor[\s\S]*?onBlur=\{/);
   });
 });
